requirements.txt
asgiref==3.4.1
autopep8==1.6.0
Django==3.2.9
django-filter==21.1
djangorestframework==3.12.4
djangorestframework-simplejwt==5.0.0
numpy==1.21.4
pandas==1.3.4
pycodestyle==2.8.0
PyJWT==2.3.0
python-dateutil==2.8.2
pytz==2021.3
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
